feat(sdk): Extract SlidingSync cache functions into their own module.
